(#104) Implement IrcTransport.ircTransportEntry
