+++ src/Main.hs
@@ -3,42 +3,28 @@ module Main where
 import           Bot
 import           Control.Concurrent
 import           Control.Concurrent.STM
-import           Control.Exception
-import           Control.Monad
 import           Control.Monad.Free
 import           Data.Foldable
-import           Data.Ini
 import qualified Data.Text as T
 import           Data.Time
-import           Data.Traversable
 import qualified Database.SQLite.Simple as SQLite
 import           Effect
 import           Events()
 import           Hookup
 import           Irc.Commands ( ircPong
-                              , ircNick
-                              , ircPass
-                              , ircJoin
                               , ircPrivmsg
                               )
 import           Irc.Identifier (idText)
-import           Irc.Message (IrcMsg(Ping, Privmsg), cookIrcMsg)
-import           Irc.RawIrcMsg (RawIrcMsg, parseRawIrcMsg, asUtf8, renderRawIrcMsg)
+import           Irc.Message (IrcMsg(Ping, Privmsg))
 import           Irc.UserInfo (userNick)
 import           IrcTransport
 import           Network.HTTP.Simple
 import qualified SqliteEntityPersistence as SEP
 import           System.CPUTime
 import           System.Environment
-import           Text.Printf
 
 -- TODO(#15): utilize rate limits
 -- See https://github.com/glguy/irc-core/blob/6dd03dfed4affe6ae8cdd63ede68c88d70af9aac/bot/src/Main.hs#L32
-
-data Config = Config { configNick :: T.Text
-                     , configPass :: T.Text
-                     , configChannel :: T.Text
-                     } deriving Show
 
 data EffectState =
     EffectState { esConfig :: Config
@@ -46,59 +32,6 @@ data EffectState =
                 , esSqliteConn :: SQLite.Connection
                 , esTimeouts :: [(Integer, Effect ())]
                 }
-
-maxIrcMessage :: Int
-maxIrcMessage = 512
-
-config :: T.Text -> T.Text -> T.Text -> Config
-config nick password channel =
-    Config { configNick = nick
-           , configPass = password
-           , configChannel = T.pack $ printf "#%s" channel
-           }
-
-configFromFile :: FilePath -> IO Config
-configFromFile filePath =
-    do ini <- readIniFile filePath
-       let lookupParam section key = ini >>= lookupValue (T.pack section) (T.pack key)
-       let nick = lookupParam "User" "nick"
-       let password = lookupParam "User" "password"
-       let channel = lookupParam "User" "channel"
-       either (ioError . userError) return $ liftM3 config nick password channel
-
-twitchConnectionParams :: ConnectionParams
-twitchConnectionParams =
-    ConnectionParams { cpHost = "irc.chat.twitch.tv"
-                     , cpPort = 443
-                     , cpTls = Just TlsParams { tpClientCertificate = Nothing
-                                              , tpClientPrivateKey = Nothing
-                                              , tpServerCertificate = Nothing
-                                              , tpCipherSuite = "HIGH"
-                                              , tpInsecure = False
-                                              }
-                     , cpSocks = Nothing
-                     }
-
-withConnection :: ConnectionParams -> (Connection -> IO a) -> IO a
-withConnection params body =
-    bracket (connect params) close body
-
-authorize :: Config -> Connection -> IO ()
-authorize conf conn =
-    do sendMsg conn (ircPass $ configPass conf)
-       sendMsg conn (ircNick $ configNick conf)
-       sendMsg conn (ircJoin (configChannel conf) Nothing)
-
-readIrcLine :: Connection -> IO (Maybe IrcMsg)
-readIrcLine conn =
-    do mb <- recvLine conn maxIrcMessage
-       for mb $ \xs ->
-           case parseRawIrcMsg (asUtf8 xs) of
-             Just msg -> return $! cookIrcMsg msg
-             Nothing -> fail "Server sent invalid message!"
-
-sendMsg :: Connection -> RawIrcMsg -> IO ()
-sendMsg conn msg = send conn (renderRawIrcMsg msg)
 
 applyEffect :: EffectState -> Effect () -> IO EffectState
 applyEffect effectState (Pure _) = return effectState
@@ -189,7 +122,10 @@ singleThreadedMain configPath databasePath =
 
 -- TODO(#105): Main.logicEntry is not implemented
 logicEntry :: IncomingQueue -> OutcomingQueue -> FilePath -> IO ()
-logicEntry _ _ _ = putStrLn "Not implemented yet. See https://github.com/tsoding/HyperNerd/issues/105"
+logicEntry incoming outcoming configPath =
+    do msg <- atomically $ tryReadTQueue incoming
+       maybe (return ()) print msg
+       logicEntry incoming outcoming configPath
 
 multiThreadedMain :: FilePath -> FilePath -> IO ()
 multiThreadedMain configPath _ =
